fix filename mixups
